created seperate file for product api, and imported/exported to/from .test.js
