+++ test/product-api.test.js
@@ -1,23 +1,7 @@
+import productApi from '../src/product-api.js';
 const test = QUnit.test;
 QUnit.module('product api');
 
-//function
-const productApi = {
-    save(product) {
-        //serialize JSON
-        const json = JSON.stringify(product);
-        //save to local storage
-        localStorage.setItem('product', json);
-    },
-    get() {
-        //get from local storage
-        const json = localStorage.getItem('product');
-        //deserialize
-        const product = JSON.parse(json);
-        //return
-        return product;
-    }
-};
 
 test('creates round-trip product', (assert) => {
     // Arrange
